selenium-tests/requirements.txt
pytest==8.3.3
selenium==4.25.0
webdriver-manager==4.0.2
pytest-html==4.1.1
python-dotenv==1.0.1
